Add pg_dump process duration to logfile entries
Modified logfile entry output, so that the time it took to perform the database dump is also recorded (reported in seconds).

Since the backup operation is effectively “silent” to the user and since Resolve databases can tend to get very large after awhile (esp. if the script is run as a cron job or LaunchAgent on macOS), I find it informative to know how long it took the pg_dump process to complete each backup operation.